src/cron/scraping/requirements.txt
pika
python-dotenv
psycopg2-binary
